rename src/duey => src/duey-engine to show where the code actually lives
